Refactor calendar and dashboard

+++ app/controllers/schedules_controller.rb
@@ -6,12 +6,20 @@ class SchedulesController < ApplicationController
 	end
 
 	def assign
-		new_availability = current_teacher.availabilities.find_or_initialize_by(day: params[:col].to_i, start_time: (params[:row].to_i + 8), end_time:(params[:row].to_i + 9))
-		if new_availability.new_record?
-			new_availability.save
-		else
-			new_availability.destroy
-		end
+		size = params[:hours_ids].size
+    unless (params[:days_ids].blank? and params[:hours_ids].blank?)
+			size.times do |i|
+				if current_teacher.can_edit?
+					new_availability = current_teacher.availabilities.find_or_initialize_by(day: params[:days_ids][i].to_i, start_time: (params[:hours_ids][i].to_i + 8), end_time:(params[:hours_ids][i].to_i + 9))
+					if new_availability.new_record?
+						new_availability.save
+					else
+						new_availability.destroy
+					end
+	      end
+	    end
+    end
+    redirect_to calendar_schedules_path
 	end
 
 	def save_calendar
